+++ src/node-project.ts
@@ -8,8 +8,7 @@ import { Jest, JestOptions } from './jest';
 import { License } from './license';
 import { NodePackage, NodePackageManager, NodePackageOptions } from './node-package';
 import { Project, ProjectOptions } from './project';
-import { Publisher } from './publisher';
-import { Release, ReleaseProjectOptions } from './release';
+import { Release, ReleaseProjectOptions, Publisher } from './release';
 import { Task } from './tasks';
 import { UpgradeDependencies, UpgradeDependenciesOptions, UpgradeDependenciesSchedule } from './upgrade-dependencies';
 import { Version } from './version';
