+++ src/release/index.ts
@@ -1,1 +1,2 @@
 export * from './release';
+export * from './publisher';
